+++ src/main/java/com/zerobase/babdeusilbun/repository/MeetingRepository.java
@@ -2,12 +2,13 @@ package com.zerobase.babdeusilbun.repository;
 
 import com.zerobase.babdeusilbun.domain.Meeting;
 import com.zerobase.babdeusilbun.domain.User;
+import com.zerobase.babdeusilbun.repository.custom.CustomMeetingRepository;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface MeetingRepository extends JpaRepository<Meeting, Long> {
+public interface MeetingRepository extends JpaRepository<Meeting, Long>, CustomMeetingRepository {
 
   @Query("select m from meeting m "
       + "where m.leader = :leader "
