+++ src/test/java/com/zerobase/babdeusilbun/meeting/service/MeetingServiceTest.java
@@ -24,20 +24,16 @@ import com.zerobase.babdeusilbun.domain.Purchase;
 import com.zerobase.babdeusilbun.domain.Store;
 import com.zerobase.babdeusilbun.domain.User;
 import com.zerobase.babdeusilbun.dto.DeliveryAddressDto;
-import com.zerobase.babdeusilbun.dto.MeetingDto;
 import com.zerobase.babdeusilbun.dto.MetAddressDto;
 import com.zerobase.babdeusilbun.enums.PurchaseStatus;
 import com.zerobase.babdeusilbun.enums.PurchaseType;
 import com.zerobase.babdeusilbun.exception.CustomException;
-import com.zerobase.babdeusilbun.exception.ErrorCode;
 import com.zerobase.babdeusilbun.meeting.dto.MeetingRequest;
 import com.zerobase.babdeusilbun.meeting.dto.MeetingRequest.Create;
 import com.zerobase.babdeusilbun.meeting.scheduler.MeetingScheduler;
-import com.zerobase.babdeusilbun.meeting.scheduler.MeetingSchedulerService;
 import com.zerobase.babdeusilbun.meeting.service.impl.MeetingServiceImpl;
-import com.zerobase.babdeusilbun.repository.MeetingQueryRepository;
+import com.zerobase.babdeusilbun.repository.custom.impl.CustomMeetingRepositoryImpl;
 import com.zerobase.babdeusilbun.repository.MeetingRepository;
-import com.zerobase.babdeusilbun.repository.PurchasePaymentRepository;
 import com.zerobase.babdeusilbun.repository.PurchaseRepository;
 import com.zerobase.babdeusilbun.repository.StoreImageRepository;
 import com.zerobase.babdeusilbun.repository.StoreRepository;
@@ -72,7 +68,7 @@ public class MeetingServiceTest {
   private MeetingRepository meetingRepository;
 
   @Mock
-  private MeetingQueryRepository meetingQueryRepository;
+  private CustomMeetingRepositoryImpl meetingQueryRepository;
 
   @Mock
   private StoreImageRepository storeImageRepository;
